requirements.txt
colorlog
python-dotenv
beautifulsoup4
llama-cpp-python
requests
aiohttp
fastapi
uvicorn[standard]
omegaconf
pydantic
ray
filelock
